refactor: create index for config variables
