+++ database/getDB.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const config = require('../config/config');
+const config = require('../config');
 
 const getDB = () => {
   const fileContents = fs.readFileSync(config.dbFilePath);
