+++ database/saveDB.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const config = require('../config/config');
+const config = require('../config');
 const saveDB = (db) => {
   fs.writeFileSync(config.dbFilePath, JSON.stringify(db, null, 2));
 };
