+++ routes/authRoutes.js
@@ -4,7 +4,6 @@ const { login, register } = require('../controllers/authController');
 const router = express.Router();
 
 router.post('/register', register);
-
 router.post('/login', login);
 
 module.exports = router;
