+++ utils/encryptPassword.js
@@ -1,5 +1,5 @@
 const bcrypt = require('bcrypt');
-const config = require('../config/config');
+const config = require('../config');
 
 const encryptPassword = (password) => {
   return bcrypt.hashSync(password, config.SALT);
